+++ simple-stack-example-mvp/src/main/java/com/zhuinden/simplestackdemoexamplemvp/presentation/paths/addoredittask/AddOrEditTaskPresenter.java
@@ -108,8 +108,7 @@ public class AddOrEditTaskPresenter
             messageQueue.pushMessageTo(addOrEditTaskKey.parent(), new TasksCoordinator.SavedSuccessfullyMessage());
             backstack.goBack();
         } else {
-            backstack.setHistory(HistoryBuilder.from(backstack.getHistory()).removeUntil(TasksKey.create()).build(),
-                    StateChange.Direction.BACKWARD);
+            backstack.setHistory(HistoryBuilder.from(backstack.getHistory()).removeUntil(TasksKey.create()).build(), StateChange.BACKWARD);
         }
     }
 }